A simple local demo run for our user to test out Avro2TF:
1. Add a new set of data with more user and movie features, from 100K data
2. Structure the data into train, validate and test
3. After running it, it will generate corresponding train, validate, and test tensor data for TF Keras model to consume
4. Restructure the TF2 notebook
