--- avro2tf/src/demo/config/avro2tf_config_movielens.json
@@ -1,143 +1,0 @@
-{
-  "features": [
-    {
-      "inputFeatureInfo": {
-        "columnExpr": "userId"
-      },
-      "outputTensorInfo": {
-        "name": "userId",
-        "dtype": "long",
-        "shape": [
-          -1
-        ]
-      }
-    },
-    {
-      "inputFeatureInfo": {
-        "columnExpr": "movieId",
-        "transformConfig": {
-          "hashInfo": {
-            "hashBucketSize": 1000,
-            "numHashFunctions": 4
-          }
-        }
-      },
-      "outputTensorInfo": {
-        "name": "movieId_hashed",
-        "dtype": "long",
-        "shape": [
-          4
-        ]
-      }
-    },
-    {
-      "inputFeatureInfo": {
-        "columnExpr": "genreFeatures"
-      },
-      "outputTensorInfo": {
-        "name": "genreFeatures",
-        "dtype": "float",
-        "shape": [],
-        "isSparse" : true
-      }
-    },
-    {
-      "inputFeatureInfo": {
-        "columnExpr": "genreFeatures.term"
-      },
-      "outputTensorInfo": {
-        "name": "genreFeatures_term",
-        "dtype": "long",
-        "shape": [
-          -1
-        ]
-      }
-    },
-    {
-      "inputFeatureInfo": {
-        "columnConfig": {
-          "genreFeatures": {
-            "whitelist": [
-              "Genre"
-            ]
-          },
-          "movieLatentFactorFeatures": {
-            "blacklist": [
-              "0"
-            ]
-          }
-        },
-        "transformConfig": {
-          "hashInfo": {
-            "hashBucketSize": 100,
-            "combiner": "AVG"
-          }
-        }
-      },
-      "outputTensorInfo": {
-        "name": "genreFeatures_movieLatentFactorFeatures",
-        "dtype": "float",
-        "shape": [],
-        "isSparse" : true
-      }
-    }
-  ],
-  "labels": [
-    {
-      "inputFeatureInfo": {
-        "columnExpr": "response"
-      },
-      "outputTensorInfo": {
-        "name": "response",
-        "dtype": "double",
-        "shape": []
-      }
-    },
-    {
-      "inputFeatureInfo": {
-        "columnExpr": "movieId",
-        "transformConfig": {
-          "hashInfo": {
-            "hashBucketSize": 1000,
-            "numHashFunctions": 4
-          }
-        }
-      },
-      "outputTensorInfo": {
-        "name": "movieId_hashed_response",
-        "dtype": "long",
-        "shape": [
-          4
-        ]
-      }
-    },
-    {
-      "inputFeatureInfo": {
-        "columnConfig": {
-          "genreFeatures": {
-            "whitelist": [
-              "Genre"
-            ]
-          },
-          "movieLatentFactorFeatures": {
-            "blacklist": [
-              "0"
-            ]
-          }
-        },
-        "transformConfig": {
-          "hashInfo": {
-            "hashBucketSize": 100,
-            "combiner": "AVG"
-          }
-        }
-      },
-      "outputTensorInfo": {
-        "name": "genreFeatures_movieLatentFactorFeatures_response",
-        "dtype": "float",
-        "shape": [],
-        "isSparse" : true
-      }
-    }
-  ]
-}
